% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain.EDR.R
\name{obtain.EDR}
\alias{obtain.EDR}
\title{Title Obtain predicted EDR}
\usage{
obtain.EDR(result)
}
\arguments{
\item{result}{input result from the function Estimate.EDR.from.pilot}
}
\value{
a list of results, including EDR, FDR etc
}
\description{
Title Obtain predicted EDR
}
\examples{

EDR<-obtain.EDR(result=power.result)


}
