% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.R
\name{plot3d}
\alias{plot3d}
\title{3D plot of EDR}
\usage{
plot3d(x)
}
\arguments{
\item{x}{a MethylSeqDesign object}
}
\value{
a 3D plot of EDR, x axls is the sample size, y axls is the propotion of subsampling, z axls is the EDR.
}
\description{
3D plot of EDR
}
