% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContour.R
\name{plotContour}
\alias{plotContour}
\title{Contour plot of EDR}
\usage{
plotContour(x)
}
\arguments{
\item{x}{a MethylSeqDesign object}
}
\value{
a contour plot of EDR, x axls is the sample size and y axls is the propotion of subsampling.
}
\description{
Contour plot of EDR
}
