% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designOptim.R
\name{designOptim}
\alias{designOptim}
\title{Optimize N and R}
\usage{
designOptim(EDR.result, pilot.R, prop, target.N, targetEDR = NULL,
  budget = NULL, price_per_lane = 2000, fix_price_per_sample = 100,
  depth_per_lane = 150)
}
\arguments{
\item{EDR.result}{a MethylSeqDesign object}

\item{pilot.R}{sequencing depth of pilot data in millions}

\item{prop}{a numeric vector of propotion of subsampling}

\item{target.N}{a numeric vector or a matrix of target sample size. If it is a matrix, the number of columns is 2, and the number of rows
equals the number of target sample size.}

\item{targetEDR}{the target EDR want to achieve}

\item{budget}{the budget limit}

\item{price_per_lane}{the price per lane for Methyl-Seq}

\item{fix_price_per_sample}{the fixed part of the price per sample}
}
\value{
a list of the optimal EDR/budget, and the corresponding sample size in two groups and sequencing depth per sample in million reads.
}
\description{
Optimize N and R
}
