% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.EDR.from.pilot.R
\name{Estimate.EDR.from.pilot}
\alias{Estimate.EDR.from.pilot}
\title{Title Estimate EDR based on p values}
\usage{
Estimate.EDR.from.pilot(p.values, model, N0, target.N, thresh.p = 0.005,
  FDR = 0.05, M = 10, method = "CBUM+CDD")
}
\arguments{
\item{p.values}{resulting p values from DMR analysis}

\item{model}{results from DMR analysis}

\item{N0}{pilot sample size per group}

\item{target.N}{target sample size per group}

\item{thresh.p}{threshold for CBUM mixture model fitting, default is set to 0.005}

\item{FDR}{False discovery rate, by default is set to 0.05}

\item{M}{Number of parametric bootstrap, default is 10}

\item{method}{by defualt 'CBUM+CDD'}
}
\value{
a list of estimation result
}
\description{
Title Estimate EDR based on p values
}
