% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.EDR.from.pilot.R
\name{Estimate.EDR.from.pilot}
\alias{Estimate.EDR.from.pilot}
\title{Title Estimate EDR based on p values}
\usage{
Estimate.EDR.from.pilot(res, N0, target.N, thresh.p = 0.005,
  FDR = 0.05, M = 10)
}
\arguments{
\item{res}{result object from DMR analysis}

\item{N0}{a numeric vector of length 2 containing sample size of the pilot data, the number of each element is 
the sample size for each group. If the two groups have the same sample size, N0 can be a single numeric value.}

\item{target.N}{target sample size per group. It is a matrix of two columns, each row contains the group sizes
for one target sample size.}

\item{thresh.p}{threshold for CBUM mixture model fitting, default is set to 0.005}

\item{FDR}{the level of False discovery rate to control, by default is set to 0.05}

\item{M}{Number of parametric bootstrap, default is 10}
}
\value{
a list of 3 element. The first element is a matrix of EDR, the second element is a matrix of declared
Number of DMR, the third element is a matrix of empirical FDR.
}
\description{
Title Estimate EDR based on p values
}
